Move manager to services
